+++ packages/AI/RemoteBrowser/Providers/SelfHost/CHANGELOG.md
@@ -1,5 +1,42 @@
 # @memberjunction/remote-browser-selfhost
 
+## 6.0.0
+
+### Patch Changes
+
+- Updated dependencies [a2670a9]
+  - @memberjunction/core@6.0.0
+  - @memberjunction/computer-use@6.0.0
+  - @memberjunction/remote-browser-base@6.0.0
+  - @memberjunction/remote-browser-cdp@6.0.0
+  - @memberjunction/global@6.0.0
+
+## 5.51.0
+
+### Patch Changes
+
+- Updated dependencies [a8fc549]
+  - @memberjunction/core@5.51.0
+  - @memberjunction/computer-use@5.51.0
+  - @memberjunction/remote-browser-base@5.51.0
+  - @memberjunction/remote-browser-cdp@5.51.0
+  - @memberjunction/global@5.51.0
+
+## 5.50.0
+
+### Patch Changes
+
+- Updated dependencies [623dfc5]
+- Updated dependencies [ce6374c]
+- Updated dependencies [deb02b4]
+- Updated dependencies [0ba33b3]
+- Updated dependencies [dd04a24]
+  - @memberjunction/core@5.50.0
+  - @memberjunction/remote-browser-base@5.50.0
+  - @memberjunction/computer-use@5.50.0
+  - @memberjunction/remote-browser-cdp@5.50.0
+  - @memberjunction/global@5.50.0
+
 ## 5.49.0
 
 ### Patch Changes