+++ packages/ThemeEngine/CHANGELOG.md
@@ -1,3 +1,9 @@
 # @memberjunction/theme-engine
 
+## 6.0.0
+
+## 5.51.0
+
+## 5.50.0
+
 ## 5.49.0